updated to GO SDK 1.6.387

+++ alicloud/service_alicloud_cen.go
@@ -200,7 +200,7 @@ func (s *CenService) WaitForCenBandwidthPackage(id string, status Status, bandwi
 				return WrapError(err)
 			}
 		}
-		if object.Status == string(status) && object.Bandwidth == bandwidth {
+		if object.Status == string(status) && object.Bandwidth == int64(bandwidth) {
 			return nil
 		}
 		if time.Now().After(deadline) {
